requirements.txt
requests==2.32.3
python-dotenv==1.0.1
qrcode==7.4.2
websockets==13.1